Module STM

A state machine framework supporting both sequential and parallel model-based testing.

Sub-module of various utility functions

include module type of Util

val repeat : int -> ( 'a -> bool ) -> 'a -> bool

repeat num prop iterates a property prop num times, returning false if just one of the iterations returns false. This is handy if the property outcome is non-determistic, for example, if it depends on scheduling.

val set_ci_printing : unit -> unit

This is a hack to delimit the printed output to the CI logs. Expect this to change or disappear.

exception Timeout

exception raised by prop_timeout and fork_prop_with_timeout.

val prop_timeout : int -> ( 'a -> 'b ) -> 'a -> 'b

prop_timeout s prop returns a property working as prop that times out and raises Timeout after s seconds.

val fork_prop_with_timeout : int -> ( 'a -> bool ) -> 'a -> bool

fork_prop_with_timeout s prop tests a property in a separate process and times out and raises Timeout after s seconds, like prop_timeout s prop. This is handy if the tested code can segfault or loop infinitely.

val print_triple_vertical : ?fig_indent:int -> ?res_width:int -> ?center_prefix:bool -> ( 'a -> string ) -> ('a list * 'a list * 'a list) -> string

print_triple_vertical pr (xs,ys,zs) returns a string representing a parallel trace, with xs printed first, and then ys and zs printed in parallel. Optional fig_indent indicates how many spaces it should be indented (default: 10 spaces). Optional res_width specifies the reserved width for printing each list entry (default: 20 chars). Optional center_prefix centers the sequential prefix if true (the default) and otherwise left-adjust it.

val protect : ( 'a -> 'b ) -> 'a -> ( 'b, exn ) Stdlib.result

protect f turns an exception throwing function into a result returning function.

val pp_exn : Stdlib.Format.formatter -> exn -> unit

Format-based exception pretty printer

val show_exn : ( Stdlib.Format.formatter -> ( Stdlib.Format.formatter -> exn -> unit ) -> unit ) -> string

Format-based exception to-string function

val equal_exn : exn -> exn -> bool

equality function for comparing exceptions

type 'a ty = ..

Extensible type to represent result values

type ty +=

| Unit : unit ty

| Bool : bool ty

| Char : char ty

| Int : int ty

| Int32 : int32 ty

| Int64 : int64 ty

| Float : float ty

| String : string ty

| Bytes : bytes ty

| Exn : exn ty

| Option : 'a ty -> 'a option ty

| Result : 'a ty * 'b ty -> ( 'a, 'b ) Stdlib.result ty

| List : 'a ty -> 'a list ty

| Array : 'a ty -> 'a array ty

| Seq : 'a ty -> 'a Stdlib.Seq.t ty

A range of constructors to represent built-in types

type 'a ty_show = 'a ty * ( 'a -> string )

Combinator type to represent an OCaml type along with an associated to_string function

val unit : unit ty_show

Combinator to represent the unit type

val bool : bool ty_show

Combinator to represent the bool type

val char : char ty_show

Combinator to represent the char type

val int : int ty_show

Combinator to represent the int type

val int32 : int32 ty_show

Combinator to represent the int32 type

val int64 : int64 ty_show

Combinator to represent the int64 type

val float : float ty_show

Combinator to represent the float type

val string : string ty_show

Combinator to represent the string type

val bytes : bytes ty_show

Combinator to represent the bytes type

val option : 'a ty_show -> 'a option ty_show

option t builds a t option type representation

val exn : exn ty_show

Combinator to represent the exception type

val result : 'a ty_show -> 'b ty_show -> ( 'a, 'b ) Stdlib.Result.t ty_show

result a b builds an (a,b) Result.t type representation

val list : 'a ty_show -> 'a list ty_show

list t builds a t list type representation

val array : 'a ty_show -> 'a array ty_show

array t builds a t array type representation

val seq : 'a ty_show -> 'a Stdlib.Seq.t ty_show

seq t builds a t Seq.t type representation

type res =

| Res : 'a ty_show * 'a -> res

val show_res : res -> string

module type StmSpec = sig ... end

The specification of a state machine.

module Make (Spec : StmSpec) : sig ... end

A functor to build model-based tests from a state-machine specification module